requirements.txt
streamlit
pillow
python-dotenv
replicate
requests
